lab1_rest/requirements.txt
coreapi==2.1.1
Django==1.10.4
django-cleanup==0.4.2
django-rest-swagger==0.3.10
django-user-agents==0.3.0
djangorestframework==3.5.3
drf-nested-routers==0.11.1
itypes==1.1.0
openapi-codec==1.2.1
Pillow==3.4.2
pycrypto==2.6.1
PyYAML==3.12
requests==2.12.3
simplejson==3.10.0
six==1.10.0
ua-parser==0.7.2
uritemplate==3.0.0
user-agents==1.0.1
